+++ games/tetris.py
@@ -19,35 +19,15 @@
 
 from random import randrange
 
+import copy
 import logging
 import pygame
-import os
 from pygame.locals import *
+from pygame.locals import USEREVENT
 from pygame.color import *
 
-import signal
-def handler(signum, frame):
-    pass
-
-signal.signal(signal.SIGHUP, handler)
-
-if os.getenv('TETRIS_EMU') != "TRUE":
-    os.putenv('SDL_FBDEV', '/dev/fb0')
-    os.putenv('SDL_VIDEODRIVER', 'fbcons')
-    os.putenv('SDL_NOMOUSE', '1')
-    os.putenv('SDL_NOSOUND', '1')
-    import lights
-    import buttons
-else:
-    import nolights as lights
-    import nobuttons as buttons
 
 log = logging.getLogger(__name__)
-logging.basicConfig(level=logging.DEBUG)
-
-board_width = 10
-board_height_hidden = 2
-board_height = 20 + board_height_hidden
 
 BUTTONEVENT = USEREVENT+1
 
@@ -68,50 +48,45 @@ shapes = [
 ]
 
 
-def new_board():
-    board = []
-    # XXX replace with clever map
-    for y in range(board_height):
-        board.append([])
-        for x in range(board_width):
-            board[y].append(0)
-    return board
-
 def shape_width(shape):
     return(len(shape[0]))
+
 
 def shape_height(shape):
     return(len(shape))
 
+
 def drop_shape(board, shape, xpos, ypos):
     """Drop a shape into a board"""
+    # This feels ugly, get rid of the deep copy
+    # when this moves into the board class
+    this_board = copy.deepcopy(board)
     if len(shape) > 0:
-        this_board = []
+        this_pixels = []
         shape_x_max = xpos + shape_width(shape)
         shape_y_max = ypos + shape_height(shape)
-        for board_y, board_line in enumerate(board):
+        for board_y, board_line in enumerate(board['pixels']):
             line = []
             for board_x, cell in enumerate(board_line):
                 if board_x >= xpos and board_x < shape_x_max and \
                         board_y >= ypos and board_y < shape_y_max:
-                    log.debug("shape_pos: y:%s, x:%s", board_y - ypos, board_x - xpos)
-                    line.append(cell + shape[board_y - ypos][board_x - xpos])
+                    y = board_y - ypos
+                    x = board_x - xpos
+                    log.debug("drop_shape, shape_pos: y:%s, x:%s", y, x)
+                    line.append(cell + shape[y][x])
                 else:
                     line.append(cell)
-            this_board.append(line)
-        return this_board
-    else:
-        return board
-
-light_board = lights.Board()
+            this_pixels.append(line)
+        this_board['pixels'] = this_pixels
+    return this_board
 
 
-def show_board(board, shape=None, shape_x=0, shape_y=0):
+def show_board_and_shape(board, light_board, shape=None, shape_x=0, shape_y=0):
     """Display the board"""
     if shape is None:
         shape = []
-    for line in drop_shape(board, shape, shape_x, shape_y):
-        print(line)
+    for line in drop_shape(board, shape, shape_x, shape_y)['pixels']:
+        log.debug("show_board_and_shape, line: %s", line)
     light_board.show_board(drop_shape(board, shape, shape_x, shape_y))
 
 
@@ -122,14 +97,15 @@ def rotate_shape(shape):
 
 
 def remove_row(board, row):
-    del board[row]
-    return board + [[0 for i in range(board_width)]]
+    log.debug("remove_row, Removing row: %s", row)
+    del board['pixels'][row]
+    board['pixels'] += [[0 for i in range(board['width'])]]
 
 
 class Tetris():
     """The testris game"""
-    def __init__(self):
-        self.board = new_board()
+    def __init__(self, board, light_board, table_buttons):
+        self.board = board
         self.running = False
         self.clock = pygame.time.Clock()
         self.drop_interval = 1000  # 1 second
@@ -141,12 +117,14 @@ class Tetris():
         self.ended = False
         self.score = 0
         self.lines_dropped = 0
+        self.light_board = light_board
+        self.buttons = table_buttons
 
     def new_shape(self):
         """Select a new shape and position it"""
         self.shape = shapes[randrange(len(shapes))]
-        self.shape_pos_x = randrange(board_width - shape_width(self.shape))
-        self.shape_pos_y = board_height - shape_height(self.shape) - board_height_hidden
+        self.shape_pos_x = randrange(self.board['width'] - shape_width(self.shape))
+        self.shape_pos_y = self.board['height'] - shape_height(self.shape) - self.board['height_hidden']
         if self.check_collisions():
             log.debug("End of Game...")
             self.ended = True
@@ -161,34 +139,35 @@ class Tetris():
         """Move the shape down one"""
         self.last_drop_time = pygame.time.get_ticks()
         self.shape_pos_y -= 1
-        print(self.shape_pos_y)
+        log.debug("make_drop, shape_pos_y: %s", self.shape_pos_y)
         hit = False
         if self.check_collisions():
-            log.debug('Collided %s', self.shape_pos_y)
+            log.debug('make_drop, Collided shape_pos_y: %s', self.shape_pos_y)
             self.shape_pos_y += 1
             self.stick_shape()
             self.new_shape()
             hit = True
-        show_board(self.board, self.shape, self.shape_pos_x, self.shape_pos_y)
+        show_board_and_shape(self.board, self.light_board, self.shape, self.shape_pos_x, self.shape_pos_y)
         return hit
 
     def stick_shape(self):
         """Stick the shape on the board"""
+        log.debug("stick_shape, Sticking shape")
         for shape_y, shape_row in enumerate(self.shape):
             for shape_x, shape_cell in enumerate(shape_row):
                 try:
-                    self.board[shape_y + self.shape_pos_y][shape_x + self.shape_pos_x] += shape_cell
+                    self.board['pixels'][shape_y + self.shape_pos_y][shape_x + self.shape_pos_x] += shape_cell
                 except:
-                    log.error("Dropping shape:x-%s, y-%s, %s, %s", shape_x, shape_y, self.shape_pos_x, self.shape_pos_y)
+                    log.error("stick_shape, Dropping shape:x-%s, y-%s, %s, %s", shape_x, shape_y, self.shape_pos_x, self.shape_pos_y)
 
         self.check_full_lines()
 
     def check_full_lines(self):
         lines_closed=0
         while True:
-            for i, row in enumerate(self.board[:-1]):
+            for i, row in enumerate(self.board['pixels'][:-1]):
                 if 0 not in row:
-                    self.board = remove_row(self.board, i)
+                    remove_row(self.board, i)
                     lines_closed += 1
                     self.lines_dropped += 1
                     break
@@ -205,19 +184,19 @@ class Tetris():
 
     def check_collisions(self):
         """Check the block is not bumping into anything"""
-        print("Checking Collisions")
+        log.info("check_collisions, Checking Collisions")
         for shape_y, shape_row in enumerate(self.shape):
             for shape_x, shape_cell in enumerate(shape_row):
                 x_pos = self.shape_pos_x + shape_x
                 y_pos = self.shape_pos_y + shape_y
-                log.debug('x_check: %s, y_check: %s', x_pos, y_pos)
+                log.debug('check_collisions, x_pos: %s, y_pos: %s', x_pos, y_pos)
                 try:
                     if y_pos < 0 or x_pos < 0:
                         return True
-                    if self.board[y_pos][x_pos] and shape_cell:
+                    if self.board['pixels'][y_pos][x_pos] and shape_cell:
                         return True
                 except IndexError:
-                    log.debug("index error:%s %s", x_pos, y_pos)
+                    log.debug("check_collisions, index error:%s %s", x_pos, y_pos)
                     return True
         return False
 
@@ -226,29 +205,29 @@ class Tetris():
         # log.debug("Buttons pressed %s", new_values)
         for x, val in enumerate(new_values):
             if not val and self.button_state[x]:
-                log.debug("Button state: %s", self.button_state)
+                log.debug("buttons_pressed, Button state: %s", self.button_state)
                 pressed.append(x)
             self.button_state[x] = val
         return pressed
 
     def button_event(self, button):
-        if button == 4:
+        if button == 1: # 4
             self.move_left()
-        elif button == 7:
+        elif button == 3: # 7
             self.move_right()
-        elif button == 5:
+        elif button == 2: # 5
             self.rotate()
-        elif button == 6:
+        elif button == 0: # 6
             self.quick_drop()
 
     def move_right(self):
         """Move the shape right"""
-        if self.shape_pos_x + len(self.shape[0]) < board_width:
+        if self.shape_pos_x + len(self.shape[0]) < self.board['width']:
             self.shape_pos_x += 1
             if self.check_collisions():
                 self.shape_pos_x -= 1
             else:
-                show_board(self.board, self.shape, self.shape_pos_x, self.shape_pos_y)
+                show_board_and_shape(self.board, self.light_board, self.shape, self.shape_pos_x, self.shape_pos_y)
 
     def move_left(self):
         """Move the shape right"""
@@ -257,17 +236,17 @@ class Tetris():
             if self.check_collisions():
                 self.shape_pos_x += 1
             else:
-                show_board(self.board, self.shape, self.shape_pos_x, self.shape_pos_y)
+                show_board_and_shape(self.board, self.light_board, self.shape, self.shape_pos_x, self.shape_pos_y)
 
     def rotate(self):
         self.shape = rotate_shape(self.shape)
-        log.debug("Shape rotating")
+        log.info("rotate, Shape rotating")
         if self.check_collisions():
             self.shape = rotate_shape(self.shape)
             self.shape = rotate_shape(self.shape)
             self.shape = rotate_shape(self.shape)
         else:
-            show_board(self.board, self.shape, self.shape_pos_x, self.shape_pos_y)
+            show_board_and_shape(self.board, self.light_board, self.shape, self.shape_pos_x, self.shape_pos_y)
 
     def quick_drop(self):
         while not self.make_drop():
@@ -275,20 +254,20 @@ class Tetris():
 
     def run(self):
         """Run the game"""
-        print("Run the game")
+        log.info("run, Run the game")
         pygame.display.init()
         pygame.init()
         # button_event = pygame.event.Event(BUTTONEVENT, message="Button Pressed", button_values=[])
         try:
-            background_sound = pygame.mixer.Sound("./tetris.ogg")
-            background_sound.play(loops = -1)
-            self.row_sound = pygame.mixer.Sound("./jump.wav")
+            background_sound = pygame.mixer.Sound("./sounds/tetris.ogg")
+            background_sound.play( loops=-1 )
+            self.row_sound = pygame.mixer.Sound("./sounds/jump.wav")
         except:
-            pass
-       
-        controls = buttons.Buttons()
-        show_board(self.board)
-        print(self.shape)
+            log.debug("run, Failed to play sound: tetris.ogg")
+
+        controls = self.buttons
+        show_board_and_shape(self.board, self.light_board)
+        log.debug("run, shape: %s", self.shape)
         self.running = True
         pygame.time.set_timer(USEREVENT, 100)  # every 100 miliseconds
         pygame.time.set_timer(USEREVENT+1, 10)  # every 10 milliseconds
@@ -301,7 +280,7 @@ class Tetris():
                     self.make_drop()
             if event.type == USEREVENT+1:
                 for button in self.buttons_pressed(controls.get_buttons()):
-                    log.debug("Button pressed %s", button)
+                    log.debug("run, Button pressed %s", button)
                     self.button_event(button)
             if event.type == KEYDOWN:
                 if event.key == pygame.K_LEFT:
@@ -324,6 +303,7 @@ def main():
     """Start the main game"""
     game = Tetris()
     game.run()
+
 
 if __name__ == "__main__":
     main()